// src/apollo/types/types.ts
export enum ChainType {
  SOLANA = "SOLANA",
  EVM = "EVM",
}

export interface IProjectPower {
  powerRank: number;
  totalPower?: number;
  round: number;
}

export interface IAdminUser {
  id?: string;
  email?: string;
  name?: string;
  walletAddress?: string;
  avatar?: string;
}

export interface IEstimatedMatching {
  allProjectsSum: number;
  matchingPool: number;
  projectDonationsSqrtRootSum: number;
}

export interface IAnchorContractData {
  address: Address;
  isActive: boolean;
}

export interface IProject {
  id: string;
  title?: string;
  balance?: number;
  image?: string;
  slug: string;
  creationDate?: string;
  adminUserId?: number;
  description?: string;
  descriptionSummary?: string;
  addresses?: IWalletAddress[];
  impactLocation?: string;
  qualityScore?: number;
  verified?: boolean;
  verificationStatus?: EProjectVerificationStatus;
  listed?: boolean | null;
  categories: ICategory[];
  reaction?: IReaction;
  adminUser: IAdminUser;
  donations: {
    id?: string;
  }[];
  totalDonations?: number;
  totalProjectUpdates?: number;
  status: {
    id?: string;
    name?: EProjectStatus;
  };
  updatedAt: string;
  latestUpdateCreationDate?: string;
  organization?: {
    name: string;
    label: string;
    supportCustomTokens: boolean;
    disableRecurringDonations?: boolean;
  };
  projectVerificationForm?: IProjectVerification;
  projectPower: IProjectPower;
  verificationFormStatus?: EVerificationStatus;
  projectFuturePower: IProjectPower;
  givbackFactor?: number;
  countUniqueDonors?: number;
  countUniqueDonorsForActiveQfRound?: number;
  estimatedMatching: IEstimatedMatching;
  sumDonationValueUsdForActiveQfRound?: number;
  anchorContracts: IAnchorContractData[];
  socialMedia: IProjectSocialMedia[];
}
// ...
export interface IWalletAddress {
  address?: string;
  isRecipient?: boolean;
  networkId?: number;
  chainType?: ChainType;
}

export interface IProjectEdition {
  id?: string;
  title?: string;
  image?: string;
  description?: string;
  addresses?: IWalletAddress[];
  impactLocation?: string;
  categories: ICategory[];
  adminUser: {
    walletAddress?: string;
  };
  status: {
    name?: string;
  };
  slug: string;
  anchorContracts?: IAnchorContractData[];
  socialMedia: IProjectSocialMedia[];
}

export enum EProjectSocialMediaType {
  FACEBOOK = "FACEBOOK",
  X = "X",
  INSTAGRAM = "INSTAGRAM",
  YOUTUBE = "YOUTUBE",
  LINKEDIN = "LINKEDIN",
  REDDIT = "REDDIT",
  DISCORD = "DISCORD",
  FARCASTER = "FARCASTER",
  LENS = "LENS",
  WEBSITE = "WEBSITE",
  TELEGRAM = "TELEGRAM",
  GITHUB = "GITHUB",
}

export interface IProjectSocialMedia {
  type: EProjectSocialMediaType;
  link: string;
}

export interface IProjectCreation {
  title: string;
  description: string;
  impactLocation?: any;
  categories: any;
  organisationId: number;
  walletAddress?: string;
  addresses?: IWalletAddress[];
  image?: string;
  isDraft?: boolean;
  socialMedia?: IProjectSocialMedia[];
}

export interface IUser {
  id?: string;
  firstName?: string;
  lastName?: string;
  name?: string;
  email?: string;
  avatar?: string;
  walletAddress?: string;
  url?: string;
  location?: string;
  userId?: string;
  totalDonated?: number;
  totalReceived?: number;
  projectsCount?: number;
  donationsCount?: number;
  likedProjectsCount?: number;
  boostedProjectsCount?: number;
  isSignedIn: boolean;
  wasReferred?: boolean;
  isReferrer?: boolean;
  chainvineId?: string;
}

export interface IPassportInfo {
  passportScore: number;
  passportStamps: any;
  activeQFMBDScore?: number | null;
}

export interface IUserWithPassport extends IUser, IPassportInfo {}

export interface IReaction {
  id: string;
  userId: string;
}
// ...
export interface ICategory {
  name: string;
  value?: string;
  isActive?: boolean;
  canUseOnFrontend?: boolean;
  mainCategory?: Pick<IMainCategory, "title">;
}
// ...
export interface IMainCategory {
  title: string;
  description: string;
  banner: string;
  slug: string;
  categories: ICategory[];
  selected?: boolean;
}
export interface IProjectRegistry {
  isNonProfitOrganization?: boolean;
  organizationCountry?: string;
  organizationWebsite?: string;
  organizationDescription?: string;
  organizationName?: string;
  attachments?: string[];
}

export interface IProjectContact {
  name: string;
  url: string;
}

export interface IProjectMilestones {
  foundationDate?: string;
  mission?: string;
  achievedMilestones?: string;
  achievedMilestonesProofs?: string[];
  problem?: string;
  plans?: string;
  impact?: string;
}

export interface IProjectManagingFunds {
  description: string;
  relatedAddresses: IAddress[];
}

export interface ISocialProfile {
  id: string;
  isVerified: boolean;
  socialNetwork: string;
  socialNetworkId: string;
  name: string;
}

export interface IPersonalInfo {
  email: string;
  walletAddress: string;
  fullName: string;
}

export interface IProjectVerification {
  id: string;
  isTermAndConditionsAccepted: boolean;
  emailConfirmationToken?: string;
  emailConfirmationSent?: boolean;
  emailConfirmationSentAt?: string;
  emailConfirmedAt?: string;
  emailConfirmed?: boolean;
  email?: string;
  projectRegistry?: IProjectRegistry;
  projectContacts?: IProjectContact[];
  milestones?: IProjectMilestones;
  managingFunds?: IProjectManagingFunds;
  emailConfirmationTokenExpiredAt?: string;
  user: IUser;
  project: IProject;
  socialProfiles?: ISocialProfile[];
  status: EVerificationStatus;
  lastStep: EVerificationSteps;
  personalInfo?: IPersonalInfo;
}

export enum EVerificationStatus {
  VERIFIED = "verified",
  DRAFT = "draft",
  SUBMITTED = "submitted",
  REJECTED = "rejected",
}

export interface IProjectVerificationUpdateInput {
  step: EVerificationSteps;
  projectVerificationId: number;
  projectRegistry?: IProjectRegistry;
  projectContacts?: IProjectContact[];
  milestones?: IProjectMilestones;
  managingFunds?: IProjectManagingFunds;
  isTermAndConditionsAccepted?: boolean;
}

export enum EVerificationSteps {
  BEFORE_START = "beforeStart",
  PERSONAL_INFO = "personalInfo",
  SOCIAL_PROFILES = "socialProfiles",
  PROJECT_REGISTRY = "projectRegistry",
  PROJECT_CONTACTS = "projectContacts",
  MANAGING_FUNDS = "managingFunds",
  IMPACT = "milestones",
  TERM_AND_CONDITION = "termAndCondition",
  SUBMIT = "submit",
}